+++ src/app/(queries)/build-filters/community-builds.ts
@@ -66,6 +66,26 @@ LEFT JOIN User on Build.createdById = User.id
 LEFT JOIN BuildReports on Build.id = BuildReports.buildId AND BuildReports.userId = ${userId}
 LEFT JOIN PaidUsers on User.id = PaidUsers.userId
 LEFT JOIN BuildTags on Build.id = BuildTags.buildId
+  LEFT JOIN (
+    SELECT 
+        BuildItems.buildId,
+        SUM(CASE WHEN BuildItems.category = 'archtype' THEN 1 ELSE 0 END) as archtypeCount,
+        SUM(CASE WHEN BuildItems.category = 'skill' THEN 1 ELSE 0 END) as skillCount,
+        SUM(CASE WHEN BuildItems.category = 'helm' THEN 1 ELSE 0 END) as helmCount,
+        SUM(CASE WHEN BuildItems.category = 'torso' THEN 1 ELSE 0 END) as torsoCount,
+        SUM(CASE WHEN BuildItems.category = 'gloves' THEN 1 ELSE 0 END) as glovesCount,
+        SUM(CASE WHEN BuildItems.category = 'legs' THEN 1 ELSE 0 END) as legsCount,
+        SUM(CASE WHEN BuildItems.category = 'relic' THEN 1 ELSE 0 END) as relicCount,
+        SUM(CASE WHEN BuildItems.category = 'relicfragment' THEN 1 ELSE 0 END) as relicfragmentCount,
+        SUM(CASE WHEN BuildItems.category = 'weapon' THEN 1 ELSE 0 END) as weaponCount,
+        SUM(CASE WHEN BuildItems.category = 'mod' THEN 1 ELSE 0 END) as modCount,
+        SUM(CASE WHEN BuildItems.category = 'mutator' THEN 1 ELSE 0 END) as mutatorCount,
+        SUM(CASE WHEN BuildItems.category = 'amulet' THEN 1 ELSE 0 END) as amuletCount,
+        SUM(CASE WHEN BuildItems.category = 'ring' THEN 1 ELSE 0 END) as ringCount,
+        SUM(CASE WHEN BuildItems.category = 'trait' THEN BuildItems.amount ELSE 0 END) as traitSum
+    FROM BuildItems
+    GROUP BY BuildItems.buildId
+  ) as ItemCounts ON Build.id = ItemCounts.buildId
 ${whereConditions}
 AND (User.displayName LIKE ${'%' + searchText + '%'} 
 OR User.name LIKE ${'%' + searchText + '%'} 
@@ -96,8 +116,29 @@ export function communityBuildsCountQuery({
   SELECT COUNT(DISTINCT Build.id) as totalBuildCount
   FROM Build
   LEFT JOIN BuildVoteCounts ON Build.id = BuildVoteCounts.buildId
+  LEFT JOIN BuildTags on Build.id = BuildTags.buildId
   LEFT JOIN User on Build.createdById = User.id
-  ${whereConditions}
+  LEFT JOIN (
+  SELECT 
+      BuildItems.buildId,
+      SUM(CASE WHEN BuildItems.category = 'archtype' THEN 1 ELSE 0 END) as archtypeCount,
+      SUM(CASE WHEN BuildItems.category = 'skill' THEN 1 ELSE 0 END) as skillCount,
+      SUM(CASE WHEN BuildItems.category = 'helm' THEN 1 ELSE 0 END) as helmCount,
+      SUM(CASE WHEN BuildItems.category = 'torso' THEN 1 ELSE 0 END) as torsoCount,
+      SUM(CASE WHEN BuildItems.category = 'gloves' THEN 1 ELSE 0 END) as glovesCount,
+      SUM(CASE WHEN BuildItems.category = 'legs' THEN 1 ELSE 0 END) as legsCount,
+      SUM(CASE WHEN BuildItems.category = 'relic' THEN 1 ELSE 0 END) as relicCount,
+      SUM(CASE WHEN BuildItems.category = 'relicfragment' THEN 1 ELSE 0 END) as relicfragmentCount,
+      SUM(CASE WHEN BuildItems.category = 'weapon' THEN 1 ELSE 0 END) as weaponCount,
+      SUM(CASE WHEN BuildItems.category = 'mod' THEN 1 ELSE 0 END) as modCount,
+      SUM(CASE WHEN BuildItems.category = 'mutator' THEN 1 ELSE 0 END) as mutatorCount,
+      SUM(CASE WHEN BuildItems.category = 'amulet' THEN 1 ELSE 0 END) as amuletCount,
+      SUM(CASE WHEN BuildItems.category = 'ring' THEN 1 ELSE 0 END) as ringCount,
+      SUM(CASE WHEN BuildItems.category = 'trait' THEN BuildItems.amount ELSE 0 END) as traitSum
+  FROM BuildItems
+  GROUP BY BuildItems.buildId
+) as ItemCounts ON Build.id = ItemCounts.buildId
+${whereConditions}
 AND (User.displayName LIKE ${'%' + searchText + '%'} 
 OR User.name LIKE ${'%' + searchText + '%'} 
 OR Build.name LIKE ${'%' + searchText + '%'}