+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## 2024-06-06
 
+### Added
+
+- You can now filter builds by a specific Relic.
+- You can now filter builds to only show "Quality" builds. A quality build has 100+ character descriptions, at least one build tag, and all item slots full (except consumables).
+
 ### Changed
 
 - Updated logo and profile image in navbar to be larger.